Final version with interpreters and sync between solvers/interp
