+++ fr.lip6.move.gal.itscl.concurrent/src/fr/lip6/move/gal/itscl/interprete/IInterpreteObservable.java
@@ -1,0 +1,8 @@
+package fr.lip6.move.gal.itscl.interprete;
+
+import java.util.concurrent.Callable;
+
+public interface IInterpreteObservable extends  Callable<Boolean> {
+	public void addThInterprete(Thread o);
+
+}
