--- fr.lip6.move.gal.itscl.concurrent/src/fr/lip6/move/gal/itscl/modele/IInterpreteObservable.java
@@ -1,7 +1,0 @@
-package fr.lip6.move.gal.itscl.modele;
-
-
-public interface IInterpreteObservable extends Ender, Runnable {
-	public void launchInterprete(IListener o);
-
-}
