+++ fr.lip6.move.gal.itscl.concurrent/src/fr/lip6/move/gal/itscl/modele/IListener.java
@@ -1,7 +1,6 @@
 package fr.lip6.move.gal.itscl.modele;
 
-import java.util.concurrent.Callable;
 
-public interface IListener extends Callable<Object>{
+public interface IListener {
 	
 }
