--- fr.lip6.move.gal.itscl.concurrent/src/fr/lip6/move/gal/itscl/modele/Problem.java
@@ -1,7 +1,0 @@
-package fr.lip6.move.gal.itscl.modele;
-
-
-public abstract class Problem implements IProblem {
-
-
-}
